Compute the correlation function at each redshift
Calculate the correlation length and slope.
Confirm the fit of the correlation length and slope.
Determine the percentage of satellites at each redshift.
Evaluate the percentage of satellites in red and blue galaxies.
Check that the total number of galaxies in each dataset is correct.
Compare the mass of blue and red satellite galaxies with a histogram.
Recalculate all figures involving red and blue galaxies to correct my error.
Compare the masses of all blue and red galaxies using the recalculated data.
Investigate z=1.00 by computing the correlation function and comparing masses
Find the "threshold" minimum stellar mass of each galaxy selection (Figs. 12 and 13).